fix error templates

+++ src/app/models/user.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 from hashlib import md5
 from time import time
 from flask import current_app
-from flask_login import UserMixin
+from flask_login import UserMixin, AnonymousUserMixin
 from app.extensions import bcrypt, login
 from app.models.post import Post
 from app.models.notification import Notification
@@ -37,6 +37,7 @@ class User(UserMixin, db.Model):
     bio = db.Column(db.String(140))
     profile_img_url = db.Column(db.String)
     last_user_notification_read_time = db.Column(db.DateTime)
+    role_id = db.Column(db.Integer, db.ForeignKey('roles.id'))
 
     posts = db.relationship(
         'Post',
@@ -49,40 +50,42 @@ class User(UserMixin, db.Model):
         primaryjoin=(followers.c.follower_id == id),
         secondaryjoin=(followers.c.followed_id == id),
         backref=db.backref('followers', lazy='joined'),
-        # backref=db.backref('followers', lazy='dynamic'),
-        # lazy='dynamic'
+        lazy='dynamic'
     )
     likes = db.relationship(
         'Post',
         secondary=likes,
         primaryjoin=(likes.c.user_id == id),
         backref=db.backref('likes', lazy='joined'),
-        # lazy='dynamic'
+        lazy='dynamic'
     )
     comments = db.relationship(
         'Comment',
         backref='author',
-        # lazy='dynamic'
+        lazy='dynamic'
     )
     notifications = db.relationship(
         'Notification',
         backref='user',
-        # lazy='dynamic'
+        lazy='dynamic'
     )
     user_notification_sent = db.relationship(
         'UserNotification',
         foreign_keys='UserNotification.sender_id',
         backref='author',
-        # lazy='dynamic'
+        lazy='dynamic'
     )
     user_notification_received = db.relationship(
         'UserNotification',
         foreign_keys='UserNotification.recipient_id',
         backref='recipient',
-        # lazy='dynamic'
+        lazy='dynamic'
     )
 
-    def __init__(self, username, email, password=None, **kwargs):
+    def __init__(self, username, email, role=None, password=None, **kwargs):
+        super(User, self).__init__(**kwargs)
+        if role:
+            self.role = role
         db.Model.__init__(self, username=username, email=email, **kwargs)
         if password:
             self.set_password(password)
@@ -94,6 +97,15 @@ class User(UserMixin, db.Model):
 
     def check_password(self, value):
         return bcrypt.check_password_hash(self.password, value)
+
+    def can(self, permissions):
+        return self.role.permissions >= permissions
+
+    def is_demo_admin(self):
+        return self.can(Permission.DEMO_ADMINISTER)
+
+    def is_admin(self):
+        return self.can(Permission.ADMINISTER)
 
     def avatar(self, size):
         digest = md5(self.email.lower().encode('utf-8')).hexdigest()
@@ -149,10 +161,12 @@ class User(UserMixin, db.Model):
             UserNotification.timestamp > last_read_time).count()
 
     def add_notification(self, name, data):
-        self.notifications.filter_by(name=name).delete()
-        n = Notification(name=name, payload_json=json.dumps(data), user=self)
-        db.session.add(n)
-        return n
+        if self.notifications:
+            self.notifications.filter_by(name=name).delete()
+            n = Notification(
+                name=name, payload_json=json.dumps(data), user=self)
+            db.session.add(n)
+            return n
 
     def get_my_following(self):
         my_following = self.query.join(
@@ -169,6 +183,63 @@ class User(UserMixin, db.Model):
             .filter(followers.c.followed_id == self.id)
         print(my_followers, sys.stdout)
         return my_followers
+
+    @property
+    def follower_count(self):
+        if self.followers:
+            return len(self.followers)
+        return 0
+
+    @property
+    def followed_count(self):
+        if self.followed and self.followed.all():
+            return len(self.followed.all())
+        return 0
+
+
+class AnonymousUser(AnonymousUserMixin):
+    def can(self, _):
+        return False
+
+    def is_admin(self):
+        return False
+
+    def is_demo_admin(self):
+        return False
+
+
+class Permission:
+    GENERAL = 0
+    DEMO_ADMINISTER = 1
+    ADMINISTER = 2
+
+
+class Role(db.Model):
+    __tablename__ = 'roles'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(64), unique=True)
+    index = db.Column(db.String(64))
+    default = db.Column(db.Boolean, default=False, index=True)
+    permissions = db.Column(db.Integer)
+    users = db.relationship('User', backref='role', lazy='dynamic')
+
+    @staticmethod
+    def insert_roles():
+        roles = {
+            'User': (Permission.GENERAL, 'main', True),
+            'DemoAdministrator': (Permission.DEMO_ADMINISTER, 'demo_admin', False),
+            'Administrator': (Permission.ADMINISTER, 'admin', False)
+        }
+        for r in roles:
+            role = Role.query.filter_by(name=r).first()
+            if role is None:
+                role = Role(name=r)
+            role.permissions = roles[r][0]
+            role.index = roles[r][1]
+            role.default = roles[r][2]
+            db.session.add(role)
+        db.session.commit()
 
 
 @login.user_loader
